Fix use of double math functions for float arguments.
(cherry picked from commit a5b150cded44d26faaa5426ea952292f7a699805)

+++ addons/primitives/line_soft.c
@@ -89,8 +89,8 @@ static void get_interpolation_parameters(int start_x, int start_y, ALLEGRO_VERTE
       *param /= lensq;
    }
    
-   dx = fabs(dx);
-   dy = fabs(dy);
+   dx = fabsf(dx);
+   dy = fabsf(dy);
    if (dx > dy)
       *minor_delta_param = dx / lensq;
    else
